requirements.txt
pillow==8.0.1
APScheduler==3.6.3
gspread==3.6.0
dropbox==10.7.0